feat: add recorder module
